Add cheesy way to "align" channels
Channel alignment needed for json output:
Because of the current design, there is no way to check if an sql
query is failing before actually beginning to produce items. That
means that if there are many queries, the first items of those queries
must be produced (and error checked) before any json can be written
in response, because if any of them are failing, the whole request fails.